requirements.txt
APScheduler==3.6.3
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.4
cryptography==3.3.1
decorator==4.4.2
flake8==3.8.4
mccabe==0.6.1
mypy==0.790
mypy-extensions==0.4.3
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
python-dotenv==0.15.0
python-telegram-bot==13.1
pytz==2020.4
six==1.15.0
toml==0.10.2
tornado==6.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
